Migrated test_util_py35 to test_util
